requirements.txt
# Working Requirements for Python 3.11 (Windows Compatible)
# Updated to include all listed libraries, algorithms, and ML models

# ============================================
# Flask Web Framework & Extensions
# ============================================
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
werkzeug==3.0.0

# ============================================
# Database
# ============================================
pymongo==4.6.0
bson==0.5.10

# ============================================
# Configuration & Environment
# ============================================
python-dotenv==1.0.0
email-validator==2.1.0

# ============================================
# File Upload and Processing
# ============================================
pandas>=2.2.0
python-docx==0.8.11
PyPDF2==3.0.1
striprtf==0.0.26

# ============================================
# HTTP Requests
# ============================================
requests==2.32.5

# ============================================
# Web Content Extraction
# ============================================
feedparser
trafilatura
newspaper3k
beautifulsoup4

# ============================================
# Natural Language Processing (NLP) Core
# ============================================
nltk>=3.9
spacy
transformers
torch

# ============================================
# Sentiment Analysis
# ============================================
textblob==0.19.0
vaderSentiment==3.3.2

# ============================================
# Keyword Extraction
# ============================================
rake-nltk==1.0.6
yake==0.4.8

# ============================================
# Text Summarization
# ============================================
sumy==0.11.0

# ============================================
# Translation Services
# ============================================
deep-translator==1.11.4
googletrans
argostranslate

# ============================================
# Language Detection
# ============================================
langdetect==1.0.9

# ============================================
# Password Hashing
# ============================================
bcrypt

# ============================================
# Additional Dependencies
# ============================================
python-socketio==5.10.0
eventlet==0.33.3

# ============================================
# Machine Learning
# ============================================
scikit-learn
numpy
scipy

# ============================================
# Notes
# ============================================
# Install spaCy model: python -m spacy download en_core_web_sm
# Algorithms: LSA, Luhn, RAKE, YAKE, Rule-Based Matching, Hybrid Keyword + ML Classifier, Zero-Shot Classification, SHA256 Hashing, Regex Cleaning, Weighted Scoring
# ML Models: BERTweet, en_core_web_sm, facebook/bart-large-mnli, Custom ML Classifier